Move completion command to separate package
