+++ cmd/completion/cmd.go
@@ -6,7 +6,7 @@
 // The copyright notice above does not evidence any actual or intended
 // publication of such source code.
 
-package cmd
+package completion
 
 import (
 	"fmt"
@@ -35,7 +35,7 @@ func supportedShells() []string {
 	}
 }
 
-func completionCmd() *cobra.Command {
+func CompletionCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   fmt.Sprintf("completion [%s]", strings.Join(supportedShells(), "|")),
 		Short: "Generate shell completion script",
@@ -79,20 +79,20 @@ PowerShell:
 		DisableFlagsInUseLine: true,
 		Args:                  cobra.MatchAll(cobra.ExactArgs(1)),
 		ValidArgs:             supportedShells(),
-		RunE: func(_ *cobra.Command, args []string) error {
+		RunE: func(cmd *cobra.Command, args []string) error {
 			shell := Shell(args[0])
 
 			switch shell {
 			case ShellBash:
-				return RootCmd.GenBashCompletion(os.Stdout)
+				return cmd.Root().GenBashCompletion(os.Stdout)
 			case ShellZsh:
 				// Cobra v1.1.1+ supports GenZshCompletion
-				return RootCmd.GenZshCompletion(os.Stdout)
+				return cmd.Root().GenZshCompletion(os.Stdout)
 			case ShellFish:
 				// the `true` gives fish the “__fish_use_subcommand” behavior
-				return RootCmd.GenFishCompletion(os.Stdout, true)
+				return cmd.Root().GenFishCompletion(os.Stdout, true)
 			case ShellPowerShell:
-				return RootCmd.GenPowerShellCompletionWithDesc(os.Stdout)
+				return cmd.Root().GenPowerShellCompletionWithDesc(os.Stdout)
 			default:
 				return fmt.Errorf("unsupported shell %q", args[0])
 			}
